A system for booking hotels. (At runtime there are several instances (at least 3); e.g. from the tourism companies ‘HolidayInTheMountains’, ‘TravelToTheSea’ and ‘AwayFromHome’). This system forwards the booking requests regarding the hotels to be booked to the travel broker and at the end logs the results in the sense of: ‘Everything booked’, ‘Hotel X could not be booked, therefore the following reservations were cancelled: U,V,W,...’. Make sure that you use meaningful time stamps so that the status of the system can be traced if necessary.
The travel broker who coordinates the bookings of itineraries and makes the hotel bookings with the
relevant hotel services. There is exactly one travel broker at runtime. The travel broker is responsible for
booking all the hotels in a booking request. It is not specified whether the travel broker requests all
hotels (servers) at the same time or step by step. In any case, the travel broker must be able to handle it
if the booking of a hotel is not successful either due to technical or business problems!